refactor: move actions on item delete to a separate file
